# swagger.yaml
swagger: "2.0"
info:
  title: Book Store APIs
  version: "v1.0"

paths:
  /health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      produces:
        - application/json
      responses:
        200:
          description: Service is healthy
          schema:
            type: object
            properties:
              status:
                type: string
                example: "healthy"
              timestamp:
                type: string
                format: date-time
        500:
          description: Service is unhealthy

  /books:
    get:
      summary: List all books
      operationId: getBooks
      produces:
        - application/json
      responses:
        200:
          description: A list of books
          schema:
            type: array
            items:
              $ref: '#/definitions/Book'
        500:
          description: Internal Server Error
    post:
      summary: Create a new book
      operationId: createBook
      consumes:
        - application/json
      parameters:
        - in: body
          name: book
          required: true
          schema:
            $ref: '#/definitions/Book'
      responses:
        201:
          description: Book created
        400:
          description: Bad Request
        500:
          description: Internal Server Error

  /books/{id}:
    get:
      summary: Get a book by ID
      operationId: getBook
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Book details
          schema:
            $ref: '#/definitions/Book'
        404:
          description: Book not found
        500:
          description: Internal Server Error
    put:
      summary: Update a book
      operationId: updateBook
      consumes:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: book
          required: true
          schema:
            $ref: '#/definitions/Book'
      responses:
        200:
          description: Updated book
        400:
          description: Bad Request
        404:
          description: Book details not found
        500:
          description: Internal Server Error
    delete:
      summary: Delete a book
      operationId: deleteBook
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Book deleted
        404:
          description: Book details not found
        500:
          description: Internal Server Error

definitions:
  Book:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      author:
        type: string
      year:
        type: integer
  APIResponse:
    type: object
    properties:
      code:
        type: integer
        format: int64
        description: HTTP-like status code
      data:
        type: object
        description: The actual response payload (can be any type)
      success:
        type: boolean
        example: true
        description: Indicates whether the operation was successful
      error:
        type: string
        description: Indicates whether the operation was successful